# docs/swagger.yaml
definitions:
...
  models.DownloadYoutubeVideoInput:
    properties:
      youtube_url:
        type: string
    type: object
...
  models.VideoResponse:
    properties:
      url:
        type: string
    type: object
host: 127.0.0.1:3000
info:
  contact:
    name: Yasin Çakır
    url: https://www.linkedin.com/in/yasincakir26/
  title: Simple JWT Auth API With Gin Framework
  version: "1.0"
paths:
...
  /youtube/download:
    post:
      consumes:
      - application/json
      parameters:
      - description: Download Youtube Video
        in: body
        name: download
        required: true
        schema:
          $ref: '#/definitions/models.DownloadYoutubeVideoInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.VideoResponse'
      summary: Youtube Download Video
      tags:
      - youtube-downloader
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"